modified

+++ src/SLIDING_WINDOW_AND_POINTERS/CountTheNumberOfFairPairs.java
@@ -22,8 +22,8 @@ public class CountTheNumberOfFairPairs {
 
     }
 
-    private  int helper(int[] nums, int target){
-        
+    private long helper(int[] nums, int target){
+
     }
 
     /*
